refactor group selector
